[EA3-137] feature:회원에 대한 조회,수정,삭제에 대해 인증정보를 통해 하도록 변경 및 Google로그인 시점에 계정 선택하도록 변경

+++ src/main/java/grep/neogul_coder/domain/prtemplate/service/LinkService.java
@@ -20,6 +20,9 @@ public class LinkService {
 
     public void deleteByPrId(Long prId) {
         Link link = linkRepository.findByPrId(prId);
+        if(link == null){
+            return;
+        }
         link.delete();
     }
